requirements.txt
agent-framework --pre
azure-identity>=1.15.0
python-dotenv>=1.0.0
flask>=3.0.0

# Observability & Tracing for Azure AI Foundry
azure-monitor-opentelemetry>=1.6.0
azure-ai-projects>=1.0.0b9
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-semantic-conventions>=0.45b0
opentelemetry-instrumentation-flask>=0.45b0
